feat : all task complete

+++ src/main/java/ru/yandex/practicum/filmorate/model/Mpa.java
@@ -3,12 +3,11 @@ package ru.yandex.practicum.filmorate.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.NonNull;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Mpa {
-    private Long id;
+    private int id;
     private String name;
 }
