+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaRowMapper.java
@@ -13,7 +13,7 @@ public class MpaRowMapper implements RowMapper<Mpa> {
     public Mpa mapRow(ResultSet rs, int rowNum) throws SQLException {
         Mpa mpa = new Mpa();
 
-        mpa.setId(rs.getLong("id"));
+        mpa.setId(rs.getInt("id"));
         mpa.setName(rs.getString("name"));
 
         return mpa;
